refactor: extract base types into separate file
